working on new modules for host discovery
